Rename the two instance/about pages and provide an option to disable the "About EpiLink" one
